+++ src/modules/Awards/routes.js
@@ -1,18 +1,19 @@
 import express from 'express';
-import { deleteAwards, updateAwards, viewAllAwardsOfUser } from './handlers.js';
-import {  deleteAwardsValidator, fetchallAwardsValidator, updateAwardsValidator } from './helpers.js';
+import { deleteAwards, viewAllAwardsOfUser } from './handlers.js';
+import {  deleteAwardsValidator, fetchallAwardsValidator,  } from './helpers.js';
 import { AwardsHelper } from '../routeConstants.js'
 import { createAwardsController } from './createAwards/createAwardsController.js';
 import { awardsAuthMiddleware } from '../../middleware/awardsAuthMiddleware.js';
+import { editAwardsController } from './editAwards/editAwardsController.js';
 const awardsRoutes = express.Router();
 
-awardsRoutes.post(AwardsHelper.addUserAwards, awardsAuthMiddleware, async (req, res) => {
+awardsRoutes.post(AwardsHelper.addUserAwards, awardsAuthMiddleware, async (req, res,next) => {
 
-    return await createAwardsController(req, res);
+    return await createAwardsController(req, res,next);
 });
-awardsRoutes.patch(AwardsHelper.editUserAwards, updateAwardsValidator, async (req, res) => {
+awardsRoutes.patch(AwardsHelper.editUserAwards, awardsAuthMiddleware, async (req, res,next) => {
 
-    return await updateAwards(req, res);
+    return await editAwardsController(req, res,next);
 });
 awardsRoutes.post(AwardsHelper.viewUserAwards, fetchallAwardsValidator, async (req, res) => {
