+++ src/utils/Awards/awardsConstants.js
@@ -17,3 +17,4 @@ export const successMessage={
 }
 
 export const allowedFields = ['awardTitle', 'description', 'issuedBy', 'issuedDate'];
+export const allowedFieldsForEditAwards = ['awardTitle', 'description', 'issuedBy', 'issuedDate', "awardId", "approvalStatus"];
